+++ src/routes/command/find.ts
@@ -1,5 +1,5 @@
 import Joi from 'joi';
-import { ControllerGeneratorOptions } from '../../comms/controller';
+import { ControllerGeneratorOptionsWithClientAndSupplier } from '../../comms/controller';
 
 import { schema as commandSchema, Command } from '../../models/command';
 import { schema as pinGroupSchema, PinGroup } from '../../models/pin-group';
@@ -37,7 +37,7 @@ interface Response {
   rows: ResponseRow[];
 }
 
-const controllerGeneratorOptions: ControllerGeneratorOptions = {
+const controllerGeneratorOptions: ControllerGeneratorOptionsWithClientAndSupplier = {
   method: 'get',
   path: '/',
   query: tableQuerySchemaGenerator().keys({