+++ src/routes/subscription/find.ts
@@ -1,5 +1,5 @@
 import Joi from 'joi';
-import { ControllerGeneratorOptions } from '../../comms/controller';
+import { ControllerGeneratorOptionsWithoutClientOrSupplier } from '../../comms/controller';
 
 import { schema as userSubscriptionSchema, UserSubscription } from '../../models/user-subscription';
 import { schema as supplierSchema, Supplier } from '../../models/supplier';
@@ -33,7 +33,7 @@ interface Response {
   rows: ResponseRow[];
 }
 
-const controllerGeneratorOptions: ControllerGeneratorOptions = {
+const controllerGeneratorOptions: ControllerGeneratorOptionsWithoutClientOrSupplier = {
   method: 'get',
   path: '/',
   right: {}, // all logged in users